requirements.txt
chromadb
PyPDF2
openai
streamlit==0.82.0
python-dotenv